refactor(ids): Split security/ids/mod.rs into focused modules
Addresses issue #165 - Refactor security IDS module

Changes:
- Split security/ids/mod.rs (544 lines) into 4 focused files:
  * types.rs (168 lines): Type definitions (DetectionType, Severity,
    DetectionResult, AttackDetails, SourceInfo, RecommendedAction,
    GeoLocation, RequestData)
  * config.rs (74 lines): IDSConfig and IDSStats with Default impl
  * detector.rs (269 lines): IntrusionDetectionSystem implementation
  * mod.rs (61 lines): Re-exports only
- Updated examples/ids_demo.rs to match new IDSConfig field names
- Added missing IDSStats fields: total_alerts, detection_rate
- Preserved existing submodules: alerts, behavioral, network, signature

Testing:
- All 377 tests passing
- No Clippy warnings
- Public API unchanged (no breaking changes)
